+++ cut.cordova.js
@@ -1,3 +1,0 @@
-module.exports = require('./lib/main');
-module.exports.Mouse = require('./lib/mouse');
-require('./lib/loader.cordova');