+++ platform/web.js
@@ -1,0 +1,3 @@
+module.exports = require('../lib/main');
+module.exports.Mouse = require('../lib/mouse');
+require('../lib/loader.web');